fix(Block Editor): deleting blocks in certain places broke editor
Deleting a tell block situated before an if/else block would
break the editor. It seems that Mithril was trying to use the
"key" property to avoid re-rendering nodes, but as a consequence
it was getting confused. Removing the key from the nodes fixes
the issue. Theoretically this could hurt performance, but for now
this avoids any issues.