fix(Map): don't take advantage of image caching
When Phaser loads an image, it checks the 'completed' property
of the image to determine whether it was in cache. (loadImageTag
function) For some reason, in Firefox, even though this property
is true and the width and height are defined, the width/height
later get unset, resulting in an invalid image that Phaser cannot
use. Disabling this check makes Phaser wait for the browser to
report that the image is ready.

Closes #46